Add test for duckdb tableprovider
