--- bin/duckdb_test/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "duckdb_test"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-description = "Spice OSS Flight Subscriber"
-
-[dependencies]
-arrow = "49.0.0"
-arrow-flight = "49.0.0"
-clap.workspace = true
-tokio.workspace = true
-async-stream = "0.3.5"
-futures = "0.3.30"
-tracing = "0.1.40"
-tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
-duckdb_datafusion = { path = "../../crates/duckdb_datafusion" }
-datafusion = "34.0.0"
-duckdb = { version = "0.9.2", features = ["bundled", "r2d2"] }
-r2d2 = "0.8.10"
